// frontend/src/config/generated.ts
// Auto-generated by scripts/sync-onchain-config.mjs
// Do not edit manually.
export const onchainConfig = {
  chainId: 84532,
  panikExecutor: "0xA28D684B06a2711badEA447aA668b7dbb6de8557",
  lockChecker: "0xB5f730e8C2B658D3fFFdbC10536039ce4D6d22C2",
  usdc: "0x036CbD53842c5426634e7929541eC2318f3dCF7e",
} as const;